Display Shift times near the chosen employees
When selecting an employee, the time is now displayed on the right
side. This could be used to make individual adjustments (+- a few
minutes) in future.
Serious internal cleanup: Discovered lots of spaghetti and put the html
form at a more convenient place, so that in future a new form for the
time adjustments can be made